==> 0_data_preprocessing/1_get_log_return_from_mid_price_1min.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
import os

df_1min = pd.read_csv('../LOB_516_Minutely_2007_2021/data_1min.csv')

df_1min

### mid price to log return

data = pd.read_csv('./mid_price_1min.csv')

log_returns = np.log(data.iloc[:, 2:]).ffill().bfill().diff()

log_returns = log_returns.loc[data['time'] != '09:30:00']

log_returns = pd.concat([data.loc[data['time'] != '09:30:00', ['date', 'time']], log_returns], axis=1)
log_returns

log_returns.to_csv('./log_returns_1min.csv', index=False)
